Move integration tests into test_nodes where they belong
